extract more granular modules
